+++ skribenten-web/frontend/src/pages/Brevredigering/OldComponents/LetterEditor/actions/updateLetter.ts
@@ -1,7 +1,8 @@
 import produce from "immer";
 
+import type { RenderedLetter } from "~/types/brevbakerTypes";
+
 import type { Action } from "../lib/actions";
-import type { RenderedLetter } from "../lib/model/skribenten";
 import type { LetterEditorState } from "../model/state";
 
 export const updateLetter: Action<LetterEditorState, [letter: RenderedLetter]> = produce((draft, letter) => {
