+++ skribenten-web/frontend/src/pages/Brevredigering/OldComponents/LetterEditor/components/Paragraph.tsx
@@ -1,5 +1,6 @@
+import type { ParagraphBlock } from "~/types/brevbakerTypes";
+
 import type { BlockProperties } from "../BlockProperties";
-import type { ParagraphBlock } from "../lib/model/skribenten";
 import ContentGroup from "./ContentGroup";
 import styles from "./Paragraph.module.css";
