+++ skribenten-web/frontend/src/pages/Brevredigering/OldComponents/LetterEditor/components/SakspartView.tsx
@@ -1,4 +1,5 @@
-import type { Sakspart } from "../lib/model/skribenten";
+import type { Sakspart } from "~/types/brevbakerTypes";
+
 import styles from "./SakspartView.module.css";
 
 export interface SakspartViewProperties {
