--- skribenten-web/frontend/src/pages/Brevredigering/OldComponents/LetterEditor/lib/model/skribenten/brevbaker.ts
@@ -1,15 +1,0 @@
-export type Distribusjonstype = "VEDTAK" | "VIKTIG" | "ANNET";
-export type LanguageCode = "BOKMAL" | "NYNORSK" | "ENGLISH";
-
-export interface LetterMetadata {
-  readonly displayTitle: string;
-  readonly isSensitiv: boolean;
-  readonly distribusjonstype: Distribusjonstype;
-}
-
-export interface TemplateDescription {
-  readonly name: string;
-  readonly letterDataClass: string;
-  readonly languages: LanguageCode[];
-  readonly metadata: LetterMetadata;
-}
