--- skribenten-web/frontend/src/pages/Brevredigering/OldComponents/LetterEditor/lib/model/skribenten/index.ts
@@ -1,4 +1,0 @@
-export * from "./brevbaker";
-export * from "./letter";
-export * from "./sakDto";
-export * from "./skribenten";
