--- skribenten-web/frontend/src/pages/Brevredigering/OldComponents/LetterEditor/lib/model/skribenten/sakDto.ts
@@ -1,20 +1,0 @@
-export interface SakDto {
-  readonly sakId: number;
-  readonly foedselsnr: string;
-  readonly foedselsdato: string;
-  readonly sakType: SakType;
-}
-
-export type SakType =
-  | "AFP"
-  | "AFP_PRIVAT"
-  | "ALDER"
-  | "BARNEP"
-  | "FAM_PL"
-  | "GAM_YRK"
-  | "GENRL"
-  | "GJENLEV"
-  | "GRBL"
-  | "KRIGSP"
-  | "OMSORG"
-  | "UFOREP";
