+++ skribenten-web/frontend/tsconfig.node.json
@@ -2,8 +2,7 @@
   "compilerOptions": {
     "composite": true,
     "module": "ESNext",
-    "moduleResolution": "Node",
-    "allowSyntheticDefaultImports": true
+    "moduleResolution": "node",
   },
   "include": ["vite.config.ts"]
 }
